rename trainer file, optimize import
