+++ trainers/dc_gan_mnist_trainer.py
